requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
bottle==0.12.25
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.1
flask==1.1.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
multidict==6.0.4
pydantic==1.10.8
python-dotenv==0.21.1
requests==2.31.0
slack-sdk==3.21.3
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
WeRoBot==1.13.1
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
